@@ -1,4 +1,5 @@
 {
+    "type": "module",
     "private": true,
     "env": {
         "node": true,
@@ -11,34 +12,33 @@
         "watch": "vite build --mode development --watch --minify=false"
     },
     "dependencies": {
-        "@vitejs/plugin-vue": "^4.4.0",
-        "axios": "^1.5.1",
+        "@vitejs/plugin-vue": "^4.6.2",
         "fork-awesome": "^1.2.0",
-        "js-cookie": "^3.0.0",
-        "pinia": "^2.1.6",
+        "js-cookie": "^3.0.5",
+        "pinia": "^2.1.7",
         "resolve-url-loader": "^5.0.0",
         "vite-plugin-resolve-extension-vue": "^0.1.0",
-        "vue": "^3.3.4",
-        "vue-i18n": "^9.5.0",
-        "vue-loader": "^17.2.0",
-        "vue-multiselect": "^3.0.0-beta.2",
-        "vue-router": "^4.2.5"
+        "vue": "^3.4.27",
+        "vue-i18n": "^9.13.1",
+        "vue-loader": "^17.4.2",
+        "vue-multiselect": "^3.0.0",
+        "vue-router": "^4.3.2"
     },
     "devDependencies": {
-        "@tailwindcss/aspect-ratio": "^0.4.0",
-        "@tailwindcss/forms": "^0.5.6",
+        "@tailwindcss/aspect-ratio": "^0.4.2",
+        "@tailwindcss/forms": "^0.5.7",
         "@tailwindcss/line-clamp": "^0.4.4",
-        "@tailwindcss/typography": "^0.5.10",
-        "acorn": "^8.10.0",
-        "autoprefixer": "^10.4.16",
+        "@tailwindcss/typography": "^0.5.13",
         "cross-env": "^7.0.3",
-        "laravel-vite-plugin": "^0.8.1",
         "lodash": "^4.17.21",
+        "sass": "^1.77.4",
+        "sass-loader": "^14.2.1",
+        "typescript": "^5.4.5",
+        "autoprefixer": "^10.4.2",
+        "axios": "^1.6.4",
+        "laravel-vite-plugin": "^1.0.0",
         "postcss": "^8.4.31",
-        "sass": "^1.69.0",
-        "sass-loader": "^13.3.2",
-        "tailwindcss": "^3.3.3",
-        "typescript": "^5.2.2",
-        "vite": "^4.4.11"
+        "tailwindcss": "^3.1.0",
+        "vite": "^5.0"
     }
 }
